requirements.txt
fitz
pdfplumber
pytesseract
Pillow
camelot-py[cv]